+++ Assets/Scripts/TrajectoryPlanner/UI/EphysCopilot/DrivePanelHandler.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using EphysLink;
 using TMPro;
+using TrajectoryPlanner.UI.EphysCopilot;
 using UnityEngine;
 
 namespace TrajectoryPlanner.UI.AutomaticManipulatorControl
