+++ Assets/Scripts/TrajectoryPlanner/UI/EphysCopilot/ResetZeroCoordinatePanelHandler.cs
@@ -2,7 +2,7 @@ using EphysLink;
 using TMPro;
 using UnityEngine;
 
-namespace TrajectoryPlanner.UI.AutomaticManipulatorControl
+namespace TrajectoryPlanner.UI.EphysCopilot
 {
     public class ResetZeroCoordinatePanelHandler : MonoBehaviour
     {
