+++ Assets/Scripts/TrajectoryPlanner/UI/EphysCopilot/ResetZeroCoordinatePanelHandler.cs.meta
