+++ Assets/Scripts/TrajectoryPlanner/UI/EphysCopilot/trajectoryplanner.ui.automaticmanipulaatorcontrol.asmdef
